% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_covar.R
\name{approx.covar}
\alias{approx.covar}
\title{Approximate the covariance matrix via factor analysis.}
\usage{
approx.covar(returns, model)
}
\arguments{
\item{returns}{A matrix of stock returns.}

\item{model}{An object of S4 class 'facmodel' generated by
the function `fit.factor.model()`.}
}
\value{
The approximated covariance matrix.
}
\description{
Approximate the covariance matrix via factor analysis.
}
