% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_unif_port.R
\name{comp.unif.port}
\alias{comp.unif.port}
\title{Compare a portfolio to the uniform portfolio}
\usage{
comp.unif.port(test, portfolio)
}
\arguments{
\item{test}{A testing set to test a portfolio on.}

\item{portfolio}{A vector of portfolio weights.}
}
\description{
Compare a portfolio to the uniform portfolio
}
