% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ersd.r
\name{covmat}
\alias{covmat}
\title{Get covariance matrix for asset returns from the correlation matrix and standard deviations.}
\usage{
covmat(cormat, sdvec)
}
\arguments{
\item{cormat}{A correlation matrix for asset class returns.}

\item{sdvec}{A vector of standard deviations.}
}
\value{
The covariance matrix.
}
\description{
Get covariance matrix for asset returns from the correlation matrix and standard deviations.
}
\examples{
covmat(stalebrink$cormat, stalebrink$ersd$sd)
covmat(rvk$cormat, rvk$ersd$sd)
covmat(horizon10year2017$cormat, horizon10year2017$ersd$sd)
}
