% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_efrontier.r
\name{efrontier}
\alias{efrontier}
\title{Compute the minimum-variance efficient frontier over a target set of portfolio expected returns.}
\usage{
efrontier(er.target.vec, ersd, cormat, aa.lb = -1e+09, aa.ub = 1e+09)
}
\arguments{
\item{er.target.vec}{Vector of targeted expected returns}

\item{ersd}{data frame with 1 row per asset class: class (class name), er (expected return), sd (standard deviation)}

\item{cormat}{Correlation matrix for asset class returns. Assumed to be positive definite.}

\item{aa.lb}{Asset allocation lower bound. Can be scalar applied to all classes, or vector of length # of asset classes.}

\item{aa.ub}{Asset allocation upper bound. Can be scalar applied to all classes, or vector of length # of asset classes.}
}
\value{
eflist if covariance matrix is positive definite, FALSE if not.
}
\description{
\code{efrontier} returns a list with efficient frontier results, plus solution information.
}
\examples{
library("magrittr")
library("quadprog")
library("tidyverse")

# Get no-bounds and bounded efficient frontiers for the Stalebrink capital market assumptions
ef.nobound <- efrontier(seq(.00, .20, .0025), stalebrink$ersd, stalebrink$cormat)
ef.noshort <- efrontier(seq(.00, .20, .0025), stalebrink$ersd, stalebrink$cormat, 0, 1)

# create a data frame with just the data we want
efdf <- bind_rows(ef.nobound$efrontier \%>\% mutate(rule="no bounds on allocation"),
                  ef.noshort$efrontier \%>\% mutate(rule="no shorts or leverage")) \%>\%
                    filter(type=="high", !is.na(per))

# graph the unbounded and bounded frontiers, and the asset-class assumptions
ggplot() +
  geom_line(data=efdf, aes(psd, per, colour=rule)) +
    scale_x_continuous(name="Standard deviation", breaks=seq(0, .3, .025), labels = scales::percent) +
    scale_y_continuous(name="Expected return", breaks=seq(0, .2, .01), labels = scales::percent) +
    ggtitle("Efficient frontier for the Stalebrink capital market assumptions", subtitle="With and without bounds on asset allocations") +
    # now add the asset-class points
    geom_point(data=ef.nobound$ersd, aes(x=sd, y=er)) +
    geom_text(data=ef.nobound$ersd, aes(x=sd, y=er, label=class), nudge_y = +.003)

}
