% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_factor_model.R
\docType{class}
\name{facmodel-class}
\alias{facmodel-class}
\title{Factor Model S4 Class}
\description{
Factor Model S4 Class
}
\section{Slots}{

\describe{
\item{\code{loadings}}{Fitted factor loadings.}

\item{\code{scores}}{Fitted factor scores.}
}}

