% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_model_fit.R
\name{fit.factor.model.new}
\alias{fit.factor.model.new}
\title{Fit a factor model to the matrix of returns}
\usage{
fit.factor.model.new(returns, k, itr)
}
\arguments{
\item{returns}{A matrix of stock returns.}

\item{k}{The number of factors to fit.}

\item{itr}{max no iterations.}
}
\value{
An S4 object from the class 'facmodel'.
}
\description{
Fit a factor model to the matrix of returns
}
