% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_factor_model.R
\name{fit.factor.model}
\alias{fit.factor.model}
\title{Fit a factor model to the matrix of returns}
\usage{
fit.factor.model(returns, k)
}
\arguments{
\item{returns}{A matrix of stock returns.}

\item{k}{The number of factors to fit.}
}
\value{
An S4 object from the class 'facmodel'.
}
\description{
Fit a factor model to the matrix of returns
}
