% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesGPfuns.R
\name{Fit_GP}
\alias{Fit_GP}
\title{Fit Gaussian Process with squared exp kernel}
\usage{
Fit_GP(x, y, sig, sigma_n, l, xstar)
}
\arguments{
\item{x}{Input time indicies}

\item{y}{Known values for input indices}

\item{sig}{Kernel hyper parameter}

\item{sigma_n}{Noise parameter}

\item{l}{Kernel hyper parameter}

\item{xstar}{Time indices to predict values for}
}
\value{

}
\description{
Fit Gaussian Process with squared exp kernel
}
