% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{horizon10year2017}
\alias{horizon10year2017}
\title{Horizon 2017 capital market assumptions.}
\format{List with 2 elements
\describe{
\item{ersd}{expected returns and standard deviations}
\item{cormat}{correlation matrix}
}}
\source{
Horizon 2017 report
}
\usage{
horizon10year2017
}
\description{
Expected returns and standard deviations by asset class, and correlation matrix
}
\examples{
  horizon10year2017
  horizon10year2017$cormat
}
\keyword{datasets}
