% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPCA.R
\name{initialize,myPCA_object-method}
\alias{initialize,myPCA_object-method}
\title{Title}
\usage{
\S4method{initialize}{myPCA_object}(.Object, data, dimension)
}
\arguments{
\item{myPCA_object}{}
}
\value{

}
\description{
Title
}
