% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ersd.r
\name{is.PD}
\alias{is.PD}
\title{Check whether covariance or correlation matrix is positive definite.}
\usage{
is.PD(covmat)
}
\arguments{
\item{covmat}{A covariance matrix for asset class returns.}
}
\value{
TRUE if covariance matrix is positive definite, FALSE if not.
}
\description{
A covariance matrix should be positive definite or else a variance can be negative. See, for example,
  https://epublications.bond.edu.au/cgi/viewcontent.cgi?article=1078&context=ejsie
}
\details{
There are a variety of available fixes if it is not.
}
\examples{
is.PD(covmat(stalebrink$cormat, stalebrink$ersd$sd)) # this IS positive definite
is.PD(stalebrink$cormat)

is.PD(covmat(rvk$cormat, rvk$ersd$sd)) # NOT PD
is.PD(rvk$cormat)

is.PD(covmat(horizon10year2017$cormat, horizon10year2017$ersd$sd)) # IS PD
is.PD(horizon10year2017$cormat)
}
