% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesGPfuns.R
\name{Kernel}
\alias{Kernel}
\title{Produce Covariance matrix}
\usage{
Kernel(x, y, sig, l)
}
\arguments{
\item{x}{Kernel input}

\item{y}{Kernel input}

\item{sig}{Hyper parameter}

\item{l}{Hyper parameter}
}
\value{

}
\description{
Produce Covariance matrix
}
