% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ersd.r
\name{makePDcorr}
\alias{makePDcorr}
\title{Get the nearest positive definite correlation matrix to the matrix we have.}
\usage{
makePDcorr(cormat)
}
\arguments{
\item{cormat}{A covariance matrix for asset class returns.}
}
\value{
cormat if covariance matrix is positive definite, FALSE if not.
}
\description{
Get the nearest positive definite correlation matrix to the matrix we have.
}
\examples{
library("magrittr")

is.PD(rvk$cormat)
cormat2 <- makePDcorr(rvk$cormat)
is.PD(cormat2)

# compare:
rvk$cormat \%>\% round(., 2)
cormat2 \%>\% round(., 2)
(cormat2 - rvk$cormat) \%>\% round(., 2)

}
