% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optim.r
\name{minvport}
\alias{minvport}
\title{Get the minimum-variance portfolio for a target expected return.}
\usage{
minvport(er.target, ersd, cormat, aa.lb = -1e+09, aa.ub = 1e+09)
}
\arguments{
\item{er.target}{A vector of asset-class expected returns.}

\item{ersd}{A data frame with class (asset-class names), er (expected return by class), and sd (standard deviation by class)}

\item{cormat}{Correlation martrix.}

\item{aa.lb}{Asset allocation lower bounds}

\item{aa.ub}{Asset allocation upper bounds}
}
\value{
The portfolio expected return.
}
\description{
Get the minimum-variance portfolio for a target expected return.
}
\examples{
# no restrictions on asset allocation - shorting and leverage allowed
minvport(.09, stalebrink$ersd, stalebrink$cormat)$portfolio

# shorting and leverage NOT allowed:
minvport(.09, stalebrink$ersd, stalebrink$cormat, 0, 1)$portfolio

# shorting and leverage NOT allowed, 40\% upper bound on real estate:
minvport(.09, stalebrink$ersd, stalebrink$cormat, 0, c(1, 1, 1, .4, 1, 1))$portfolio

}
