% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPCA.R
\docType{class}
\name{myPCA_object-class}
\alias{myPCA_object-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{matrix.}

\item{\code{dimension}}{integer.}

\item{\code{centre}}{numeric.}

\item{\code{loadings}}{matrix.}

\item{\code{projection}}{matrix.}

\item{\code{projected}}{matrix.}
}}

