% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_portfolio.R
\name{optim.portfolio}
\alias{optim.portfolio}
\title{Optimise portolio weights via quadratic programming.}
\usage{
optim.portfolio(returns, covar, lambda)
}
\arguments{
\item{returns}{A matrix of stock returns}

\item{covar}{An approximated covariance matrix}
}
\description{
Optimise portolio weights via quadratic programming.
}
