% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ersd.r
\name{per}
\alias{per}
\title{Get a portfolio's expected return from asset-class expected returns and weights.}
\usage{
per(ervec, wts)
}
\arguments{
\item{ervec}{A vector of asset-class expected returns.}

\item{wts}{A vector of asset-allocation weights. Should sum to 1.}
}
\value{
The portfolio expected return.
}
\description{
Get a portfolio's expected return from asset-class expected returns and weights.
}
\examples{
library("magrittr")
per(stalebrink$ersd$er, c(.3, .2, .1, .15, .1, .05))
}
