% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ersd.r
\name{psd}
\alias{psd}
\title{Get a portfolio's standard deviation from the correlation matrix, standard deviations, and asset-allocation weights.}
\usage{
psd(cormat, sdvec, wts)
}
\arguments{
\item{cormat}{A correlation matrix for asset class returns.}

\item{sdvec}{A vector of standard deviations.}

\item{wts}{A vector of asset-allocation weights. Should sum to 1.}
}
\value{
The portfolio standard deviation.
}
\description{
Get a portfolio's standard deviation from the correlation matrix, standard deviations, and asset-allocation weights.
}
\examples{
psd(stalebrink$cormat, stalebrink$ersd$sd, c(.3, .2, .1, .15, .1, .05))
}
