% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_close.R
\name{retrieve.close}
\alias{retrieve.close}
\title{Retreieve the close prices between specified dates
Columns containing missing values will}
\usage{
retrieve.close(data, from, to, missing = TRUE)
}
\arguments{
\item{data}{A dataframe to be subsetted by date. The dataframe must have a column
with the name 'date'.}

\item{from}{A character string which gives the date of the earliest
entry e.g. 2018-01-01.}

\item{to}{A character string which gives the date of the latest
entry e.g. 2018-01-01.}

\item{missing}{Set equal to false if you want to retain missing values.}
}
\value{
A dataframe with observations between the specified dates.
}
\description{
Retreieve the close prices between specified dates
Columns containing missing values will
}
