% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returns}
\alias{returns}
\title{Convert close prices into returns}
\usage{
returns(data, lag, log = TRUE)
}
\arguments{
\item{data}{A datafram of close prices.}
}
\value{

}
\description{
The transpose of the matrix is returned so that this matrix can be used
with the rest of the package.
}
