% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{rvk}
\alias{rvk}
\title{RVK capital market assumptions used for the LAFPP pension plan.}
\format{List with 2 elements
\describe{
\item{ersd}{expected returns and standard deviations}
\item{cormat}{correlation matrix}
}}
\source{
RVK asset-liability studey
}
\usage{
rvk
}
\description{
Expected returns and standard deviations by asset class, and correlation matrix
}
\examples{
  rvk
  rvk$cormat
}
\keyword{datasets}
