% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesGPfuns.R
\name{squared_exp_kernel}
\alias{squared_exp_kernel}
\title{Squared Exponential Kernel Function}
\usage{
squared_exp_kernel(x, y, sig, l)
}
\arguments{
\item{x}{Kernel input}

\item{y}{Kernel input}

\item{sig}{Hyper parameter}

\item{l}{Hyper parameter}
}
\value{

}
\description{
Squared Exponential Kernel Function
}
