% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{stalebrink}
\alias{stalebrink}
\title{Stalebrink .}
\format{List with 2 elements
\describe{
\item{ersd}{expected returns and standard deviations}
\item{cormat}{correlation matrix}
}}
\source{
Stalebrink paper
}
\usage{
stalebrink
}
\description{
Expected returns and standard deviations by asset class, and correlation matrix
}
\examples{
  stalebrink
  stalebrink$cormat
}
\keyword{datasets}
