% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportVectorMachine.R
\name{SVM}
\alias{SVM}
\title{Support Vector Machine classifier}
\usage{
SVM(X, y, max_it = 10000, eta_0 = 10, alpha = 0.9, c = 0.9)
}
\arguments{
\item{X}{covariates}

\item{y}{response vector}

\item{max_it}{maximum iterations}

\item{eta_0}{initial learning rate}

\item{alpha}{learning rate decay}

\item{c}{cost parameter}
}
\value{
numeric vector of coefficient parameter estimates for y
}
\description{
Support Vector Machine classifier
}
\examples{
SVM(X = X, y = y, max_it = 1e4, eta_0 = 10, alpha = 0.9, c = 0.9)
}
